Refactored API functions AND made table red
