+++ src/app/stations/[stationCrs]/page.tsx
@@ -1,45 +1,5 @@
-import { fetchFromAPI } from "@/app/apiFetch";
-import {JourneyTable} from "@/app/components/JourneyTable";
-
-type StationDetailsResponse = {
-    location: {addressLines: string, postCode: string},
-    facilities: {fares?: {ticketOffice?: {openingTimes?: string}}}
-}
-
-type StationDescription = {
-    address: string,
-    ticketOfficeInfo: string
-};
-
-async function getStationDetails(stationCrs: string): Promise<StationDescription> {
-    const data = await fetchFromAPI<StationDetailsResponse>(`stationDetails/${stationCrs}`);
-    const address = data.location.addressLines.replaceAll("<br>", ", ") + ", " + data.location.postCode;
-    const ticketOfficeOpeningTimes : string | undefined = data.facilities.fares?.ticketOffice?.openingTimes;
-
-    const ticketOfficeInfo = ticketOfficeOpeningTimes
-        ? "Ticket office opening times: " + ticketOfficeOpeningTimes.replaceAll("<br>", ", ")
-        : "Ticket office information is unavailable";
-
-    return {
-        address: address,
-        ticketOfficeInfo: ticketOfficeInfo,
-    };
-}
-
-type StationNamesResponse = {
-    stations: {crs: string, name: string}[];
-}
-
-async function getStationName(stationCrs: string): Promise<string> {
-    const data = await fetchFromAPI<StationNamesResponse>("stations");
-    const stations = data.stations;
-    for (const station of stations) {
-        if (station.crs === stationCrs) {
-            return station.name;
-        }
-    }
-    throw new Error("Station does not exist.");
-}
+import { JourneyTable } from "@/app/components/JourneyTable";
+import { getStationDetails, getStationName } from "@/app/apiFunctions";
 
 export default async function Page({params}: {params: Promise<{ stationCrs: string }>}) {
     const { stationCrs: stationCrs } = await params;
